+++ cli/inspire/cli/commands/account/__init__.py
@@ -14,7 +14,6 @@ from inspire.cli.commands.account.use import use
 # the command object to the bare name would shadow the module it came from.
 from inspire.cli.commands.account.check import check as check_command
 from inspire.cli.commands.account.context import context as context_command
-from inspire.cli.commands.account.show import show as show_command
 
 
 @click.group()
@@ -28,7 +27,6 @@ account.add_command(use)
 account.add_command(remove)
 account.add_command(rename)
 account.add_command(current)
-account.add_command(show_command)
 account.add_command(check_command)
 account.add_command(context_command)
 account.add_command(permissions)